+++ django_openpay/webhooks.py
@@ -9,10 +9,21 @@ def verification(body):
 
 def chargeRefunded(body):
     transaction = body['transaction']
-    customer = models.Customer.objects.get(
-        openpay_id=transaction['customer_id'])
-    card = models.Card.objects.get(
-        openpay_id=transaction['card']['id'])
+    card = None
+    customer = None
+    subscription = None
+    if 'customer_id' in transaction:
+        customer = models.get_customer_model().objects.get(
+            openpay_id=transaction['customer_id'])
+    if 'subscription_id' in transaction:
+        subscription = models.Subscription.objects.get(
+            openpay_id=transaction['subscription_id'])
+    if 'card' in transaction:
+        card = models.Card.objects.get(
+            openpay_id=transaction['card']['id'])
+        if not customer and 'customer_id' in transaction['card']:
+            customer = models.get_customer_model().objects.get(
+                openpay_id=transaction['card']['customer_id'])
     charge, created = models.Charge.objects.update_or_create(
         openpay_id=transaction['id'],
         defaults={
@@ -57,6 +68,21 @@ def chargeRefunded(body):
 
 def chargeCancelled(body):
     transaction = body['transaction']
+    card = None
+    customer = None
+    subscription = None
+    if 'customer_id' in transaction:
+        customer = models.get_customer_model().objects.get(
+            openpay_id=transaction['customer_id'])
+    if 'subscription_id' in transaction:
+        subscription = models.Subscription.objects.get(
+            openpay_id=transaction['subscription_id'])
+    if 'card' in transaction:
+        card = models.Card.objects.get(
+            openpay_id=transaction['card']['id'])
+        if not customer and 'customer_id' in transaction['card']:
+            customer = models.get_customer_model().objects.get(
+                openpay_id=transaction['card']['customer_id'])
     charge, created = models.Charge.objects.update_or_create(
         openpay_id=transaction['id'],
         defaults={
@@ -73,15 +99,29 @@ def chargeCancelled(body):
             'order_id': transaction['order_id'],
             'amount': transaction['amount'],
             'currency': transaction['currency'],
-            'customer': models.Customer.objects.get(
-                openpay_id=transaction['customer_id']),
-            'card': models.Card.objects.get(
-                openpay_id=transaction['card']['id'])
+            'customer': customer,
+            'card': card,
+            'subscription': subscription
         })
 
 
 def chargeCreated(body):
     transaction = body['transaction']
+    card = None
+    customer = None
+    subscription = None
+    if 'customer_id' in transaction:
+        customer = models.get_customer_model().objects.get(
+            openpay_id=transaction['customer_id'])
+    if 'subscription_id' in transaction:
+        subscription = models.Subscription.objects.get(
+            openpay_id=transaction['subscription_id'])
+    if 'card' in transaction:
+        card = models.Card.objects.get(
+            openpay_id=transaction['card']['id'])
+        if not customer and 'customer_id' in transaction['card']:
+            customer = models.get_customer_model().objects.get(
+                openpay_id=transaction['card']['customer_id'])
     charge, created = models.Charge.objects.update_or_create(
         openpay_id=transaction['id'],
         defaults={
@@ -98,15 +138,29 @@ def chargeCreated(body):
             'order_id': transaction['order_id'],
             'amount': transaction['amount'],
             'currency': transaction['currency'],
-            'customer': models.Customer.objects.get(
-                openpay_id=transaction['customer_id']),
-            'card': models.Card.objects.get(
-                openpay_id=transaction['card']['id'])
+            'customer': customer,
+            'card': card,
+            'subscription': subscription
         })
 
 
 def chargeSucceeded(body):
     transaction = body['transaction']
+    card = None
+    customer = None
+    subscription = None
+    if 'customer_id' in transaction:
+        customer = models.get_customer_model().objects.get(
+            openpay_id=transaction['customer_id'])
+    if 'subscription_id' in transaction:
+        subscription = models.Subscription.objects.get(
+            openpay_id=transaction['subscription_id'])
+    if 'card' in transaction:
+        card = models.Card.objects.get(
+            openpay_id=transaction['card']['id'])
+        if not customer and 'customer_id' in transaction['card']:
+            customer = models.get_customer_model().objects.get(
+                openpay_id=transaction['card']['customer_id'])
     charge, created = models.Charge.objects.update_or_create(
         openpay_id=transaction['id'],
         defaults={
@@ -123,8 +177,7 @@ def chargeSucceeded(body):
             'order_id': transaction['order_id'],
             'amount': transaction['amount'],
             'currency': transaction['currency'],
-            'customer': models.Customer.objects.get(
-                openpay_id=transaction['customer_id']),
-            'card': models.Card.objects.get(
-                openpay_id=transaction['card']['id'])
+            'customer': customer,
+            'card': card,
+            'subscription': subscription
         })